Adds handling for partial dates
Dates were originally poorly handled if "2012" was provided, it would fail to
correctly generate the right date range. This adds handling to expand on the
date parameter such that "year" becomes "year 01 01", and "year month" becomes
"year month 01". This is consistent with the expectation of providing the time
range of "2010 until 2011" which should return everything after 2010, and
before 2011, that is, everything in 2010.