journal commit takes git args
journal-commit now correctly detects when untracked files have
been added and ensures that they are correctly added to the commit.
This change also updates the command to take git parameters so that
it can properly apply a git commit, as otherwise git detects it is
not in a tty, and will not prompt for commit message.