Overhauls Title/# handling
The markdown -> html transformer is fairly simple, and initially only
understood how to convert '^#+ ' into an <h4> tag. This changes it so
that '^# ' becomes H1, and '^##+ ' becomes H4. As part of the change
it also changes which plugin outputs the header, because the render
plugin is no longer restricted to a single Header tag type (H4), it
outputs the H1 tag for page titles.

Additionally this also cleans up some of the output HTML, by removing
BR tags when they have been added near H1, H4 and HR tags. And it
also escapes "<" and ">" characters that may appear in the text.

Note that this escaping makes inline Html impossible!