Switch to out-html/out-bin from build
Using a single output directory (build) for all output complicates
generation of the index because of the way that tup handles searching
directories for files. Listing all the files in a directory (via
find) causes them all to get added as dependencies by tup. While this
is generally desirable, it means that while we don't depend on
non-html files they do get included and tup complains we never said
that we depended on them.

To address this, html and non-html output have been split into their
own directories. This cleans up tup dependencies, and simplifies the
handling of HTML files.