Add Aliases
Adds alias functionality to tapestry threads. This directive is added
to the thread-page to denote that those names should be included in
that particular thread as well. These directives are not read by any
other threads (ie: when processing thread X no other threads aliases
are loaded), making them one-way.