Switch Storage.iterate to DependencyGraph.iterate
As was documented in the test: Storage.iterate only returns nodes
that exist in storage, and nodes are _not_ guaranteed to be added to
Storage unless the build has run. Whereas nodes are guaranteed to
exist in DependencyGraph no matter what (otherwise they're not real).

Moving the iteration ensures that even nodes without content are
properly iterated over.